Move CirclePaymentObserverService to a platform.
