--- platform/src/main/java/org/stellar/anchor/platform/paymentobserver/CirclePaymentObserver.java
@@ -1,3 +1,0 @@
-package org.stellar.anchor.platform.paymentobserver;
-
-public class CirclePaymentObserver {}
